docs: update README.md with new get function syntax and behavior
When using the get function you can now specify the expected return type in order to help with debugging and improve your code. The get function now also accepts types like number, boolean, function, bigint, symbol, array or date. You can use these type definitions along with the pathing to make it easier to retrieve values from an object.

The get function will now return the type specified when a value is found at that location in the object. If no value is found at that path then the result will be undefined.